Use common VertexRow component
